Revert "Bug 2042256 - Adjust fuzz for Skia m149. r=gfx-reviewers,media-playback-reviewers,nical,aosmond" for causing failures at test_favicons_conversions.js.
This reverts commit 33c3a75ca7301cb406402926a130057f6872ed58.

Revert "Bug 2042256 - Update mozbuilds for Skia m149. r=gfx-reviewers,nical"

This reverts commit a956c9b9cc24d84156c29cc33f410946725991ac.

Revert "Bug 2042256 - Fix Moz2D for Skia m149. r=gfx-reviewers,nical"

This reverts commit c276f84c019d91bfca4c54fef807989b669dbeef.

Revert "Bug 2042256 - Update Skia source files for Skia m149. r=gfx-reviewers,nical"

This reverts commit 283f89ec0096a59569b68040bc2a389082f57e52.

UltraBlame original commit: b0849b179c14b84bf2b93902d973d618cc019d25

+++ gfx/skia/skia/src/core/SkDraw_vertices.cpp
@@ -313,8 +313,8 @@ void Draw::drawVertices(const SkVertices* vertices,
                         const SkPaint& paint,
                         bool skipColorXform) const {
     SkVerticesPriv info(vertices->priv());
-    const size_t vertexCount = info.vertexCount();
-    const size_t indexCount = info.indexCount();
+    const int vertexCount = info.vertexCount();
+    const int indexCount = info.indexCount();
 
     
     if (vertexCount < 3 || (indexCount > 0 && indexCount < 3) || fRC->isEmpty()) {
